feat: authorization checks on dish and meal endpoints

+++ packages/server/src/module/dish/DishService.ts
@@ -3,7 +3,10 @@ import { Transaction, TransactionRepository } from "typeorm";
 import { Injectable } from "@nestjs/common";
 
 import { BadRequestError } from "../../error/BadRequestError";
+import { ForbiddenError } from "../../error/ForbiddenError";
 import { InvalidIdError } from "../../error/InvalidIdError";
+import { authorize } from "../../utility/authorization";
+import { Account } from "../account/Account";
 import { Foodstuff } from "../foodstuff/Foodstuff";
 import { FoodstuffRepository } from "../foodstuff/FoodstuffRepository";
 import { Meal } from "../meal/Meal";
@@ -19,6 +22,7 @@ export class DishService {
   @Transaction()
   public async create(
     dto: CreateDishDto,
+    principal: Account,
     @TransactionRepository() dishRepository?: DishRepository,
     @TransactionRepository() foodstuffRepository?: FoodstuffRepository,
     @TransactionRepository() mealRepository?: MealRepository
@@ -28,6 +32,8 @@ export class DishService {
     if (meal === undefined) {
       throw new InvalidIdError(Meal, ["mealId"]);
     }
+
+    authorize(principal, "add dish to", meal);
 
     const foodstuff = await foodstuffRepository!.findOne(dto.foodstuffId);
 
@@ -56,13 +62,16 @@ export class DishService {
   @Transaction()
   public async delete(
     dto: DeleteDishDto,
+    principal: Account,
     @TransactionRepository() dishRepository?: DishRepository
   ) {
-    const dish = await dishRepository!.findOne(dto.id);
+    const dish = await dishRepository!.findOne(dto.id, { relations: ["meal"] });
 
     if (dish === undefined) {
       throw new InvalidIdError(Dish, ["id"]);
     }
+
+    authorize(principal, "delete", dish);
 
     await dishRepository!.unlink(dish);
     await dishRepository!.remove(dish);
@@ -71,22 +80,29 @@ export class DishService {
   @Transaction()
   public async insert(
     dto: InsertDishDto,
+    principal: Account,
     @TransactionRepository() dishRepository?: DishRepository,
     @TransactionRepository() mealRepository?: MealRepository
   ) {
-    const dish = await dishRepository!.findOne(dto.id);
+    const dish = await dishRepository!.findOne(dto.id, { relations: ["meal"] });
 
     if (dish === undefined) {
       throw new InvalidIdError(Dish, ["id"]);
     }
 
-    await dishRepository!.unlink(dish);
-
     const meal = await mealRepository!.findOne(dto.mealId);
 
     if (meal === undefined) {
       throw new InvalidIdError(Meal, ["mealId"]);
+    } else if (dish.meal!.accountId !== meal.accountId) {
+      throw new ForbiddenError(
+        "Meal can only be inserted into a meal with the same owner."
+      );
     }
+
+    authorize(principal, "insert dish into", meal);
+
+    await dishRepository!.unlink(dish);
 
     const dishes = await dishRepository!.ordered(meal);
